ウォーニングを削除

+++ core/common/sys.h
@@ -149,7 +149,7 @@ public:
 };
 
 // ------------------------------------
-typedef unsigned int THREAD_HANDLE;
+typedef uintptr_t THREAD_HANDLE;
 #define THREAD_PROC int WINAPI
 #define vsnprintf _vsnprintf
